20.4-55-g4f62ae8d-0ubuntu1 (patches unapplied)
Imported using git-ubuntu import.

+++ tests/integration_tests/clouds.py
@@ -6,7 +6,7 @@ from pycloudlib import EC2, GCE, Azure, OCI, LXDContainer, LXDVirtualMachine
 from pycloudlib.lxd.instance import LXDInstance
 
 import cloudinit
-from cloudinit.subp import subp
+from cloudinit.subp import subp, ProcessExecutionError
 from tests.integration_tests import integration_settings
 from tests.integration_tests.instances import (
     IntegrationEc2Instance,
@@ -25,6 +25,65 @@ except ImportError:
 log = logging.getLogger('integration_testing')
 
 
+def _get_ubuntu_series() -> list:
+    """Use distro-info-data's ubuntu.csv to get a list of Ubuntu series"""
+    out = ""
+    try:
+        out, _err = subp(["ubuntu-distro-info", "-a"])
+    except ProcessExecutionError:
+        log.info(
+            "ubuntu-distro-info (from the distro-info package) must be"
+            " installed to guess Ubuntu os/release"
+        )
+    return out.splitlines()
+
+
+class ImageSpecification:
+    """A specification of an image to launch for testing.
+
+    If either of ``os`` and ``release`` are not specified, an attempt will be
+    made to infer the correct values for these on instantiation.
+
+    :param image_id:
+        The image identifier used by the rest of the codebase to launch this
+        image.
+    :param os:
+        An optional string describing the operating system this image is for
+        (e.g.  "ubuntu", "rhel", "freebsd").
+    :param release:
+        A optional string describing the operating system release (e.g.
+        "focal", "8"; the exact values here will depend on the OS).
+    """
+
+    def __init__(
+        self,
+        image_id: str,
+        os: "Optional[str]" = None,
+        release: "Optional[str]" = None,
+    ):
+        if image_id in _get_ubuntu_series():
+            if os is None:
+                os = "ubuntu"
+            if release is None:
+                release = image_id
+
+        self.image_id = image_id
+        self.os = os
+        self.release = release
+        log.info(
+            "Detected image: image_id=%s os=%s release=%s",
+            self.image_id,
+            self.os,
+            self.release,
+        )
+
+    @classmethod
+    def from_os_image(cls):
+        """Return an ImageSpecification for integration_settings.OS_IMAGE."""
+        parts = integration_settings.OS_IMAGE.split("::", 2)
+        return cls(*parts)
+
+
 class IntegrationCloud(ABC):
     datasource = None  # type: Optional[str]
     integration_instance_cls = IntegrationInstance
@@ -32,7 +91,23 @@ class IntegrationCloud(ABC):
     def __init__(self, settings=integration_settings):
         self.settings = settings
         self.cloud_instance = self._get_cloud_instance()
-        self.image_id = self._get_initial_image()
+        if settings.PUBLIC_SSH_KEY is not None:
+            # If we have a non-default key, use it.
+            self.cloud_instance.use_key(
+                settings.PUBLIC_SSH_KEY, name=settings.KEYPAIR_NAME
+            )
+        elif settings.KEYPAIR_NAME is not None:
+            # Even if we're using the default key, it may still have a
+            # different name in the clouds, so we need to set it separately.
+            self.cloud_instance.key_pair.name = settings.KEYPAIR_NAME
+        self._released_image_id = self._get_initial_image()
+        self.snapshot_id = None
+
+    @property
+    def image_id(self):
+        if self.snapshot_id:
+            return self.snapshot_id
+        return self._released_image_id
 
     def emit_settings_to_log(self) -> None:
         log.info(
@@ -50,21 +125,20 @@ class IntegrationCloud(ABC):
         raise NotImplementedError
 
     def _get_initial_image(self):
-        image_id = self.settings.OS_IMAGE
+        image = ImageSpecification.from_os_image()
         try:
-            image_id = self.cloud_instance.released_image(
-                self.settings.OS_IMAGE)
+            return self.cloud_instance.released_image(image.image_id)
         except (ValueError, IndexError):
-            pass
-        return image_id
+            return image.image_id
 
     def _perform_launch(self, launch_kwargs):
         pycloudlib_instance = self.cloud_instance.launch(**launch_kwargs)
-        pycloudlib_instance.wait(raise_on_cloudinit_failure=False)
         return pycloudlib_instance
 
-    def launch(self, user_data=None, launch_kwargs=None,
+    def launch(self, user_data=None, launch_kwargs=None, wait=True,
                settings=integration_settings):
+        if launch_kwargs is None:
+            launch_kwargs = {}
         if self.settings.EXISTING_INSTANCE_ID:
             log.info(
                 'Not launching instance due to EXISTING_INSTANCE_ID. '
@@ -73,13 +147,15 @@ class IntegrationCloud(ABC):
                 self.settings.EXISTING_INSTANCE_ID
             )
             return
+        if 'wait' in launch_kwargs:
+            raise Exception("Specify 'wait' directly to launch, "
+                            "not in 'launch_kwargs'")
         kwargs = {
             'image_id': self.image_id,
             'user_data': user_data,
             'wait': False,
         }
-        if launch_kwargs:
-            kwargs.update(launch_kwargs)
+        kwargs.update(launch_kwargs)
         log.info(
             "Launching instance with launch_kwargs:\n{}".format(
                 "\n".join("{}={}".format(*item) for item in kwargs.items())
@@ -87,9 +163,17 @@ class IntegrationCloud(ABC):
         )
 
         pycloudlib_instance = self._perform_launch(kwargs)
-
+        if wait:
+            pycloudlib_instance.wait(raise_on_cloudinit_failure=False)
         log.info('Launched instance: %s', pycloudlib_instance)
-        return self.get_instance(pycloudlib_instance, settings)
+        instance = self.get_instance(pycloudlib_instance, settings)
+        if wait:
+            # If we aren't waiting, we can't rely on command execution here
+            log.info(
+                'cloud-init version: %s',
+                instance.execute("cloud-init --version")
+            )
+        return instance
 
     def get_instance(self, cloud_instance, settings=integration_settings):
         return self.integration_instance_cls(self, cloud_instance, settings)
@@ -99,6 +183,19 @@ class IntegrationCloud(ABC):
 
     def snapshot(self, instance):
         return self.cloud_instance.snapshot(instance, clean=True)
+
+    def delete_snapshot(self):
+        if self.snapshot_id:
+            if self.settings.KEEP_IMAGE:
+                log.info(
+                    'NOT deleting snapshot image created for this testrun '
+                    'because KEEP_IMAGE is True: %s', self.snapshot_id)
+            else:
+                log.info(
+                    'Deleting snapshot image created for this testrun: %s',
+                    self.snapshot_id
+                )
+                self.cloud_instance.delete_image(self.snapshot_id)
 
 
 class Ec2Cloud(IntegrationCloud):
@@ -130,7 +227,14 @@ class AzureCloud(IntegrationCloud):
         return Azure(tag='azure-integration-test')
 
     def destroy(self):
-        self.cloud_instance.delete_resource_group()
+        if self.settings.KEEP_INSTANCE:
+            log.info(
+                'NOT deleting resource group because KEEP_INSTANCE is true '
+                'and deleting resource group would also delete instance. '
+                'Instance and resource group must both be manually deleted.'
+            )
+        else:
+            self.cloud_instance.delete_resource_group()
 
 
 class OciCloud(IntegrationCloud):
@@ -191,7 +295,6 @@ class _LxdIntegrationCloud(IntegrationCloud):
         if self.settings.CLOUD_INIT_SOURCE == 'IN_PLACE':
             self._mount_source(pycloudlib_instance)
         pycloudlib_instance.start(wait=False)
-        pycloudlib_instance.wait(raise_on_cloudinit_failure=False)
         return pycloudlib_instance
